Complete circle class implementation.

+++ Abstract Classes And Methods/Shape.cs
@@ -5,7 +5,7 @@ namespace ShapesProgram
     //Creates an abstract class Shape
     abstract class Shape
     {
-        //Creates an abstract method
-        internal protected abstract float GetArea();
+        //Creates an abstract method GetArea
+        internal protected abstract double GetArea();
     }
 }
